# backend/api_spec.yaml
openapi: "3.0.2"
info:
  title: "WaterProbe API"
  description: "API for waterprobing"
  version: "1"
paths:
...
  /Observation:
    get:
      tags:
      - Observation
      summary: "Call to get observation from database"
      operationId: getObservation
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      
      parameters: 
       - $ref: '#/components/parameters/observation'
    post:
      tags:
      - Observation
      summary: "Call to write observation to database"
      operationId: postObservation
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      
      requestBody:
        description: Optional properties represented as json for the Thing
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/observation'
...
  /sensor:
    get:
      tags: 
      - Sensor
      summary: "gets a sensor"
      operationId: getSensor
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
         $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/sensor'
        
    
    put:
      tags: 
      - Sensor
      summary: "puts a sensor"
      operationId: putSensor
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/sensor'
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sensor'
        
    post:
      tags: 
      - Sensor
      summary: "post a sensor"
      operationId: postSensor
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sensor'
        
    
    delete:
      tags: 
      - Sensor
      summary: "deletes a sensor"
      operationId: deleteSensor
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
      parameters: 
        - $ref: '#/components/parameters/sensor'
...
  /FeatureOfInterest:
    get:
      tags:
       - FeatureOfInterest
      summary: "gets a FeatureOfInterest"
      operationId: getFeatureOfInterest
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
      parameters:
      - $ref: '#/components/parameters/featureOfInterest'

    put:
      tags:
      - FeatureOfInterest
      summary: "updates a FeatureOfInterest"
      operationId: putFeatureOfInterest
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/featureOfInterest'
        
    post:
      tags:
       - FeatureOfInterest
      summary: "posting a new FeatureOfInterest"
      operationId: postFeatureOfInterest
      responses:
        201:
          $ref: '#/components/responses/201'
        404:
          $ref: '#/components/responses/404'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/featureOfInterest'
    
    delete:
      tags: 
      - FeatureOfInterest
      summary: "deletes a FeatureOfInterest"
      operationId: deleteFeatureOfInterest
      responses:
        200:
          $ref: '#/components/responses/200'
        404:
          $ref: '#/components/responses/404'
                
      parameters: 
        - $ref: '#/components/parameters/featureOfInterest'
...
components:
  responses:
    200:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    201:
      description: Successful response
      content:
        application/json:
          schema:
            title: Sample
            type: object
            properties:
              placeholder:
                type: string
                description: Placeholder description
    404:
      description: Not created response
      content:
        text/plain:
          schema:
            title: Thing not cretead
            type: string
            example:
  parameters:
    observation: 
      name: observationName
      in: query
      description: "Unique observationName"
      required: false
      schema:
        type: string
...
    sensor: 
      name: sensorName
      in: query
      description: "Unique sensorName"
      schema:
        type: string
...
    featureOfInterest:
      name: featureOfInterestName
      in: query
      description: "Unique FeatureOfInterest name"
      schema:
        type: string
...
  schemas:
...
    tmObject:
      type: object
      properties:
        time:
          type: string 
    dqElement:
      type: object
    tmInstant:
      type: object
      properties:
        time:
          type: string
    tmPeriod:
      type: object
    namedValue:
      type: object
    valueCode:
      type: object
...
    observation:
      type: object
      required:
        - phenomenonTime 
        - resultTime
        - result
        - datastreamRef
        
      properties:
        phenomenonTime:
          $ref: '#/components/schemas/tmObject'
        resultTime:
          $ref: '#/components/schemas/tmInstant'
        result:
          type: number
        resultQuality:
          type: array
          items: 
            $ref: '#/components/schemas/dqElement'
        validTime:
          type: array
          items:
            $ref: '#/components/schemas/tmPeriod'
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/namedValue'
        datastreamRef:
          type: string
        featureOfInterestRef:
          type: string
...
    sensor:
      type: object
      required:
       - name
       - description
       - encodingType
       - metaData
      
      properties:
        name:
          type: string
        description:
          type: string
        encodingType:
          $ref: '#/components/schemas/valueCode'
        metaData:
          type: object  
    featureOfInterest:
      type: object
      required: 
      - name
      - description 
      - encoding
      - feature
      properties:
        name:
          type: string
        description:
          type: string
        encoding:
          $ref: '#/components/schemas/valueCode'
        feature:
          type: object
...